+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+- [#353] Added `getRadiation` and `getRadiationRaw` to the environment detector
+
 ## [1.18.2-0.7.20r] - 2022-11-13
 
 ### Added